### 2022-10-28 #4889 @editorialbot
[PRE REVIEW]: Corekaburra: pan-genome post-processing using core gene synteny
**Submitting author:** @milnus (<a href="http://orcid.org/0000-0001-9751-9877">Magnus Jespersen</a>)
**Repository:** https://github.com/milnus/Corekaburra
**Branch with paper.md** (empty if default branch): main
**Version:** v0.0.4
