### 2022-10-30 6308771
Update paper.bib
Correct invalid doi for paper refernece
