+++ paper/paper.bib
@@ -60,7 +60,7 @@ abstract = {The development of efficient and inexpensive genome sequencing metho
     month = {07},
     abstract = "{Population-level comparisons of prokaryotic genomes must take into account the substantial differences in gene content resulting from horizontal gene transfer, gene duplication and gene loss. However, the automated annotation of prokaryotic genomes is imperfect, and errors due to fragmented assemblies, contamination, diverse gene families and mis-assemblies accumulate over the population, leading to profound consequences when analysing the set of all genes found in a species. Here, we introduce Panaroo, a graph-based pangenome clustering tool that is able to account for many of the sources of error introduced during the annotation of prokaryotic genome assemblies. Panaroo is available at https://github.com/gtonkinhill/panaroo.}",
     issn = {1474-760X},
-    doi = {doi.org/10.1186/s13059-020-02090-4},
+    doi = {10.1186/s13059-020-02090-4},
 }
 
 @article{thorpe:2018,
