### 2022-11-02 #4889 @Kevin-Mattheus-Moerman
@milnus can you check the above invalid DOI :point_up: ? You can call `@editorialbot check references` to check these again, and you can call `@editorialbot generate pdf` to update the paper.
