### 2022-11-03 #4889 @csoneson
👋🏻 @karinlag, @asafpr, @giraola - would (two of) you be interested in reviewing this submission for JOSS?

> Corekaburra: pan-genome post-processing using core gene synteny
