### 2022-11-03 #4889 @giraola
I am out-of-office right now, but I think @iferres can be interested in
having a look at it.

El jue, 3 nov 2022 a las 15:39, Charlotte Soneson ***@***.***>)
escribió:

> 👋🏻 @karinlag <https://github.com/karinlag>, @asafpr
> <https://github.com/asafpr>, @giraola <https://github.com/giraola> -
> would (two of) you be interested in reviewing this submission for JOSS?
>
> Corekaburra: pan-genome post-processing using core gene synteny
>
