### 2022-11-04 #4889 @csoneson
@evezeyl - would you be interested in reviewing this submission for JOSS (see [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) for more information about our reviewing model)?
