### 2022-11-07 milnus/Corekaburra#12 @asafpr
Installation in conda
Installation in conda requires
`conda install -c bioconda -c conda-forge corekaburra`
otherwise I got this error:

> UnsatisfiableError: The following specifications were found to be incompatible with each other:                                                                                                                                                                                                                                                                                               Output in format: Requested package -> Available versions
> The following specifications were found to be incompatible with your system:                                                                                                                                                                                                                                                           - feature:/linux-64::__glibc==2.31=0                                                                                                                                                           - python=3.9 -> libgcc-ng[version='>=7.5.0'] -> __glibc[version='>=2.17']                                                                                                                                                                                                                                                                                                                   Your installed version is: 2.31

https://github.com/openjournals/joss-reviews/issues/4910
