### 2022-11-07 #4910 @asafpr
@milnus the workflow wiki page has a broken link to gff files dir, should be https://github.com/asafpr/Corekaburra/tree/main/Example_workflow/Input_Gff_files
