### 2022-11-07 #4910 @asafpr
@milnus after `panaroo-generate-gffs` why should the file names be changed? Is there another use for that script with the original output file names? why not producing the correct file names in this script?
