### 2022-11-07 #4910 @asafpr
@milnus the file coreless_contig_accessory_gene_content.tsv is missing from output dir. Also the file core_gene_graph.gml has a typo in the wiki page (core_gene_grapgh.gml) and the last paragraph has some mixed singular/plural. In addition some of the output files are tsv while others are csv, I think they should all be the same. I didn't understand the definition of sub-segments, maybe you can explain it in more detail.
The downstream analysis in my opinion should be wrapped in scripts, at least the most common questions, it will be more user-friendly this way and I think this is the main use-case of the tool.
