### 2022-11-07 milnus/Corekaburra#15 @iferres
Contributing guidelines
For https://github.com/openjournals/joss-reviews/issues/4910 add something that satisfies "Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support".

A CONTRIBUTING.md file in the repo root should be ok, there are lots of templates in the internet. It is also a good idea to add a CODE_OF_CONDUCT.md.
