### 2022-11-13 milnus/Corekaburra#12 @milnus
The `-c conda-forge` has been added to the conda install instructions
