### 2022-11-14 9978b3e
Change output to .tsv, adjust tests, add contribution and community guide lines
