+++ CODE_OF_CONDUCT.md
@@ -1,0 +1,8 @@
+# Code of Conduct for Corekaburra users
+
+
+## Our goals
+We as maintainers of Corekaburra and part of the scientific community want to promote a fruitful, open, and inclusive forum, in which users, contributors, and others can seek help, discuss relevant topics, and fell welcome to participate. We will promote and to the best of our ability be respectful and helpful. No issue will be ignored, and we will do out best to provide people with adequate assistance to solve issues.
+
+## As a user
+As users of Corekaburra we expect you to contribute to the fruitful, open, and inclusive environment that we promote. You are not obliged to take part in discussion, issues, or code maintenance. Because of this we also ask that you do not express your opinion on any of these, if you can not live up to the standards we promote. If any expression or behaviour is deemed inappropriate by the maintainers comments will be hidden. As a user feel free to provide feedback to the community by giving solutions to issues, but be aware of not providing misinformation.
