+++ .github/workflows/code_cov_check.yml
@@ -10,9 +10,9 @@ jobs:
       OS: ubuntu-latest
       PYTHON: '3.9'
     steps:
-#      - uses: checkout@v2
-#        with:
-#          fetch-depth: ‘2’
+      - uses: actions/checkout@v3
+        with:
+          fetch-depth: 2
       - name: 'install coverage'
         - run: pip install coverage
         - run: coverage
