+++ .github/workflows/code_cov_check.yml
@@ -10,7 +10,7 @@ jobs:
       OS: ubuntu-latest
       PYTHON: '3.9'
     steps:
-      - uses: actions/checkout@v3
+      - uses: actions/checkout@v2.5.0
         with:
           fetch-depth: 2
       - name: 'install coverage'
