+++ .github/workflows/code_cov_check.yml
@@ -11,7 +11,9 @@ jobs:
       PYTHON: '3.9'
     steps:
       - name: setup
-      - uses: actions/checkout@v2.5.0
+      - uses: actions/checkout@v.2.5.0
+        with:
+          fetch-detph: 2
       - name: 'install coverage'
         - run: pip install coverage
         - run: coverage
