+++ .github/workflows/code_cov_check.yml
@@ -11,7 +11,7 @@ jobs:
       PYTHON: '3.9'
     steps:
       - name: setup
-      - uses: actions/checkout@v.2.5.0
+        uses: actions/checkout@v.2.5.0
         with:
           fetch-detph: 2
       - name: 'install coverage'
@@ -20,7 +20,7 @@ jobs:
       - name: 'Code coverage check'
         - run: coverage -m unittest unit_test/Corekaburra_test.py
       - name: 'Upload coverage report'
-      - uses: codecov/codecov-action@v3
+        uses: codecov/codecov-action@v3
         with:
           files: .coverage # optional
           flags: unittests # optional
