+++ .github/workflows/code_cov_check.yml
@@ -13,7 +13,7 @@ jobs:
       - name: setup
         uses: actions/checkout@v.2.5.0
         with:
-          fetch-detph: 2
+          fetch-detph: '2'
       - name: 'install coverage'
         - run: pip install coverage
         - run: coverage
