+++ .github/workflows/code_cov_check.yml
@@ -10,13 +10,11 @@ jobs:
       OS: ubuntu-latest
       PYTHON: '3.9'
     steps:
-      - name: setup
-        uses: actions/checkout@v.2.5.0
+      - uses: actions/checkout@v2
       - name: 'install coverage'
-        - run: pip install coverage
-        - run: coverage
+        run: pip install coverage
       - name: 'Code coverage check'
-        - run: coverage -m unittest unit_test/Corekaburra_test.py
+        run: coverage -m unittest unit_test/Corekaburra_test.py
       - name: 'Upload coverage report'
         uses: codecov/codecov-action@v3
         with:
