### 2022-11-14 eb9dc21
add the run command to coverage line
