+++ .github/workflows/code_cov_check.yml
@@ -14,7 +14,7 @@ jobs:
       - name: 'install coverage'
         run: pip install coverage
       - name: 'Code coverage check'
-        run: coverage run -m unittest unit_test/Corekaburra_test.py
+        run: coverage run -m unittest unit_tests/Corekaburra_test.py
       - name: 'Upload coverage report'
         uses: codecov/codecov-action@v3
         with:
