+++ .github/workflows/code_cov_check.yml
@@ -11,9 +11,10 @@ jobs:
       PYTHON: '3.9'
     steps:
       - uses: actions/checkout@v2
+      - name: 'install dependencies'
+        run: pip3 install -r requirements-dev.txt
       - name: 'install coverage'
         run: pip3 install coverage
-        run: pip3 install -r requirements-dev.txt
       - name: 'Code coverage check'
         run: coverage run -m unittest unit_tests/Corekaburra_test.py
       - name: 'Upload coverage report'
