+++ .github/workflows/code_cov_check.yml
@@ -10,13 +10,15 @@ jobs:
       OS: ubuntu-latest
       PYTHON: '3.9'
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v3
       - name: 'install dependencies'
         run: pip3 install -r requirements-dev.txt
       - name: 'install coverage'
         run: pip3 install coverage
+      - name: 'navigate to unittest directory'
+          run: cd unit_tests/
       - name: 'Code coverage check'
-        run: coverage run -m unittest unit_tests/Corekaburra_test.py
+        run: coverage run -m unittest Corekaburra_test.py
       - name: 'Upload coverage report'
         uses: codecov/codecov-action@v3
         with:
