+++ .github/workflows/code_cov_check.yml
@@ -16,7 +16,7 @@ jobs:
       - name: 'install coverage'
         run: pip3 install coverage
       - name: 'navigate to unittest directory'
-          run: cd unit_tests/
+        run: cd unit_tests/
       - name: 'Code coverage check'
         run: coverage run -m unittest Corekaburra_test.py
       - name: 'Upload coverage report'
