### 2022-11-15 1e4a354
Add in statement to run test for code_cov
