### 2022-11-15 53af463
Change coverage run command
