### 2022-11-15 22a0139
Change some path try/except
