+++ .github/workflows/code_cov_check.yml
@@ -24,7 +24,7 @@ jobs:
         uses: codecov/codecov-action@v3
         with:
           files: .coverage # optional
-          flags: unittests # optional
-          name: codecov-Corekaburra # optional
+#          flags: unittests # optional
+#          name: codecov-Corekaburra # optional
           fail_ci_if_error: true # optional (default = false)
           verbose: true # optional (default = false)
