+++ .github/workflows/code_cov_check.yml
@@ -18,12 +18,14 @@ jobs:
 #      - name: 'navigate to unittest directory'
 #        run: cd unit_tests/
       - name: 'Code coverage check'
-        run: coverage run -m unittest --data-file=coverage.xml unit_tests/Corekaburra_test.py
+        run: coverage run -m unittest unit_tests/Corekaburra_test.py
 #        run: coverage run -m unittest Corekaburra_test
+      - name: 'Convert coverage to xml'
+        run: coverage xml -o coverage.xml
       - name: 'Upload coverage report'
         uses: codecov/codecov-action@v3
         with:
-          files: .coverage # optional
+          files: coverage.xml # optional
 #          flags: unittests # optional
 #          name: codecov-Corekaburra # optional
           fail_ci_if_error: true # optional (default = false)
