### 2022-11-15 milnus/Corekaburra#15 @milnus
Both have not been written and included. Both are displayed in the lower right hand corner when creating a new issue.
