### 2022-11-15 #4910 @milnus
@asafpr - Response to comments:
_____
> @milnus the workflow wiki page has a broken link to gff files dir, should be

Thanks! Very thorough reading and checking of links. The dead link has now been fixed.
____
> @milnus after `panaroo-generate-gffs` why should the file names be changed? Is there another use for that script with the original output file names? why not producing the correct file names in this script?

The `panaroo-generate-gffs` script is not part of Corekaburra, but [Panaroo](https://github.com/gtonkinhill/panaroo). The file names are changed so that Corekaburra can connect the Gff file name to a column in the gene_presence_absence.csv file. The Gff files produced from the `panaroo-generate-gffs` are very unlikely to be used for any post-processing in Panaroo, as the pan-genome graph is preferred.
____
> the file coreless_contig_accessory_gene_content.tsv is missing from output dir

This file is only produced when there are at least one contig with no core genes. This has been clarified in the documentation.
____
> Also the file core_gene_graph.gml has a typo in the wiki page (core_gene_grapgh.gml) and the last paragraph has some mixed singular/plural

This has been corrected
_____
> In addition some of the output files are tsv while others are csv, I think they should all be the same.

Agreed, it was a bit silly. Now everything is converted to .tsv output.
____
> I didn't understand the definition of sub-segments, maybe you can explain it in more detail.

The paragraphs `core_segments.tsv` and `no_accessory_core_segments.tsv` have both been reworded to hopefully give a cleaner explanation around segments and sub-segments.
____
> The downstream analysis in my opinion should be wrapped in scripts, at least the most common questions, it will be more user-friendly this way and I think this is the main use-case of the tool.

I see the point, and it has been debated heavily in our group. We have produced some 'beta' scripts to do small tasks but always find it to be less flexible than using command-line tools to query output files.
Currently our view is that the amount of time for implementation and maintenance is better spend on future feature implementations, likely to detect rearrangements. If we find a query that is not easily made with command-line tools, we will definitely look to make a script.
____
> @milnus in the paper line 24 "dispite" -> "despite"

This has been corrected.
