### 2022-11-21 a7cd94a
Minor fixes to bibliography
