+++ paper/paper.bib
@@ -12,7 +12,7 @@
 
 @article{tettelin:2005,
 author = {Hervé Tettelin  and Vega Masignani  and Michael J. Cieslewicz  and Claudio Donati  and Duccio Medini  and Naomi L. Ward  and Samuel V. Angiuoli  and Jonathan Crabtree  and Amanda L. Jones  and A. Scott Durkin  and Robert T. DeBoy  and Tanja M. Davidsen  and Marirosa Mora  and Maria Scarselli  and Immaculada Margarit {y Ros}  and Jeremy D. Peterson  and Christopher R. Hauser  and Jaideep P. Sundaram  and William C. Nelson  and Ramana Madupu  and Lauren M. Brinkac  and Robert J. Dodson  and Mary J. Rosovitz  and Steven A. Sullivan  and Sean C. Daugherty  and Daniel H. Haft  and Jeremy Selengut  and Michelle L. Gwinn  and Liwei Zhou  and Nikhat Zafar  and Hoda Khouri  and Diana Radune  and George Dimitrov  and Kisha Watkins  and Kevin J. B. O'Connor  and Shannon Smith  and Teresa R. Utterback  and Owen White  and Craig E. Rubens  and Guido Grandi  and Lawrence C. Madoff  and Dennis L. Kasper  and John L. Telford  and Michael R. Wessels  and Rino Rappuoli  and Claire M. Fraser },
-title = {Genome analysis of multiple pathogenic isolates of \texttt{{S}treptococcus agalactiae}: Implications for the microbial pan-genome},
+title = {Genome analysis of multiple pathogenic isolates of \textit{{S}treptococcus agalactiae}: Implications for the microbial pan-genome},
 journal = {Proceedings of the National Academy of Sciences},
 volume = {102},
 number = {39},
@@ -126,7 +126,7 @@ abstract = {The development of efficient and inexpensive genome sequencing metho
 
 @article{welch:2002,
 author = {R. A. Welch  and V. Burland  and G. Plunkett  and P. Redford  and P. Roesch  and D. Rasko  and E. L. Buckles  and S.-R. Liou  and A. Boutin  and J. Hackett  and D. Stroud  and G. F. Mayhew  and D. J. Rose  and S. Zhou  and D. C. Schwartz  and N. T. Perna  and H. L. T. Mobley  and M. S. Donnenberg  and F. R. Blattner },
-title = {Extensive mosaic structure revealed by the complete genome sequence of uropathogenic \texttt{{E}scherichia coli}},
+title = {Extensive mosaic structure revealed by the complete genome sequence of uropathogenic \textit{{E}scherichia coli}},
 journal = {Proceedings of the National Academy of Sciences},
 volume = {99},
 number = {26},
