+++ paper/paper.bib
@@ -11,7 +11,7 @@
 }
 
 @article{tettelin:2005,
-author = {Hervé Tettelin  and Vega Masignani  and Michael J. Cieslewicz  and Claudio Donati  and Duccio Medini  and Naomi L. Ward  and Samuel V. Angiuoli  and Jonathan Crabtree  and Amanda L. Jones  and A. Scott Durkin  and Robert T. DeBoy  and Tanja M. Davidsen  and Marirosa Mora  and Maria Scarselli  and Immaculada Margarit {y Ros}  and Jeremy D. Peterson  and Christopher R. Hauser  and Jaideep P. Sundaram  and William C. Nelson  and Ramana Madupu  and Lauren M. Brinkac  and Robert J. Dodson  and Mary J. Rosovitz  and Steven A. Sullivan  and Sean C. Daugherty  and Daniel H. Haft  and Jeremy Selengut  and Michelle L. Gwinn  and Liwei Zhou  and Nikhat Zafar  and Hoda Khouri  and Diana Radune  and George Dimitrov  and Kisha Watkins  and Kevin J. B. O'Connor  and Shannon Smith  and Teresa R. Utterback  and Owen White  and Craig E. Rubens  and Guido Grandi  and Lawrence C. Madoff  and Dennis L. Kasper  and John L. Telford  and Michael R. Wessels  and Rino Rappuoli  and Claire M. Fraser },
+author = {Hervé Tettelin  and Vega Masignani  and Michael J. Cieslewicz  and Claudio Donati  and Duccio Medini  and Naomi L. Ward  and Samuel V. Angiuoli  and Jonathan Crabtree  and Amanda L. Jones  and A. Scott Durkin  and Robert T. DeBoy  and Tanja M. Davidsen  and Marirosa Mora  and Maria Scarselli  and Immaculada {Margarit y Ros}  and Jeremy D. Peterson  and Christopher R. Hauser  and Jaideep P. Sundaram  and William C. Nelson  and Ramana Madupu  and Lauren M. Brinkac  and Robert J. Dodson  and Mary J. Rosovitz  and Steven A. Sullivan  and Sean C. Daugherty  and Daniel H. Haft  and Jeremy Selengut  and Michelle L. Gwinn  and Liwei Zhou  and Nikhat Zafar  and Hoda Khouri  and Diana Radune  and George Dimitrov  and Kisha Watkins  and Kevin J. B. O'Connor  and Shannon Smith  and Teresa R. Utterback  and Owen White  and Craig E. Rubens  and Guido Grandi  and Lawrence C. Madoff  and Dennis L. Kasper  and John L. Telford  and Michael R. Wessels  and Rino Rappuoli  and Claire M. Fraser },
 title = {Genome analysis of multiple pathogenic isolates of \textit{{S}treptococcus agalactiae}: Implications for the microbial pan-genome},
 journal = {Proceedings of the National Academy of Sciences},
 volume = {102},
