### 2022-11-21 #4910 @csoneson
👋🏻 @milnus - I have gone through the submission, and made a [PR](https://github.com/milnus/Corekaburra/pull/16) with a couple of small fixes, mostly to the bibliography. Please have a look and merge if you agree. The next steps then are as follows:

- [ ] Make a tagged release of the software, and list the version tag of the archived version here.
- [ ] Check that the archival deposit (e.g., in Zenodo) has the correct metadata. In particular, the title and author list should be identical to those of the paper. You can also add the authors' ORCID.
