+++ setup.py
@@ -12,7 +12,7 @@ from standard pan-genome pipelines: Roary and Panaroo.'''
 
 setup(
     name='Corekaburra',
-    version='0.0.4',
+    version='0.0.5',
     author='Magnus Ganer Jespersen',
     author_email='magnus.ganer.j@gmail.com',
     packages=['Corekaburra'],
