### 2022-11-21 #4910 @milnus
I have created a new tagged release (v0.0.5).
Zenodo upload: doi.org/10.5281/zenodo.7344004
